backend/requirements.txt
mistralai>=0.0.17
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.23.0
scikit-learn>=1.0.0
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn>=0.23.0
requests>=2.28.0
